requirements.txt
beautifulsoup4
flask
pynacl
waitress
